+++ src/Firebase/Messaging.php
@@ -23,7 +23,6 @@ use Kreait\Firebase\Messaging\MulticastSendReport;
 use Kreait\Firebase\Messaging\RegistrationToken;
 use Kreait\Firebase\Messaging\RegistrationTokens;
 use Kreait\Firebase\Messaging\Topic;
-use Kreait\Firebase\Project\ProjectId;
 use Kreait\Firebase\Util\JSON;
 
 class Messaging implements Contract\Messaging
@@ -37,11 +36,11 @@ class Messaging implements Contract\Messaging
     /**
      * @internal
      */
-    public function __construct(ProjectId $projectId, ApiClient $messagingApiClient, AppInstanceApiClient $appInstanceApiClient)
+    public function __construct(string $projectId, ApiClient $messagingApiClient, AppInstanceApiClient $appInstanceApiClient)
     {
         $this->messagingApi = $messagingApiClient;
         $this->appInstanceApi = $appInstanceApiClient;
-        $this->projectId = $projectId->value();
+        $this->projectId = $projectId;
     }
 
     public function send($message, bool $validateOnly = false): array