+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   * `Auth\UserRecord::$customAttributes`, use `Auth\UserRecord::$customClaims` instead
   * `Factory::withEnabledDebug()`, use `Factory::withHttpDebugLogger()` instead
 * Removed deprecated/obsolete internal classes and methods)
+  * `Kreait\Firebase\Project\ProjectId`
   * `Kreait\Firebase\Value\Provider`
 
 [Unreleased]: https://github.com/kreait/firebase-php/compare/5.x...6.x